requirements.txt
# Core dependencies
openai>=1.0.0
pyyaml>=6.0
requests>=2.28.0
typing-extensions>=4.5.0
argparse>=1.4.0
concurrent-futures>=3.0.5
pathlib>=1.0.1
vllm
torch>=1.12.0